Application-specific logging.
Due to popular demand, we maintain our own logs now and have
a menu item for easy access.  Console.app (or your favorite
log viewer) is used to display the logs as to not reinvent
anything.

Rotation basically happens on restart.  It's possible to have
another facility, but there's nothing easy that comes for free.

Change-Id: Ic0b216b6a58b404a7ad6c6fa4cebffc5e3086c03
Reviewed-on: http://review.membase.org/5850
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Dustin Sallings <dustin@spy.net>