Minimal changes to get couchdbx-app to run ns_server
Much of this change is around the xcode project to allow it to build
for both Couchbase Server and Couchbase Single Server from the same
sources.

The actual app is hacked up a bit to be happy with ns_server, but the
intention is to fold this branch back into master in such a way that
everything is a build-time or runtime toggle.  Some of that is already
in.

This does mean that "xcodebuild" no longer does anything particularly
useful.  You must specify what it is you're trying to build (by target
or scheme).

    $ xcodebuild -list
    Information about project "Couchbase Server":
        Targets:
            Couchbase Single Server Zip
            Couchbase Single Server
            Couchbase ServerTests
            Couchbase Server Zip
            Couchbase Server

        Build Configurations:
            Debug
            Release

        If no build configuration is specified "Release" is used.

For example:

    $ xcodebuild -target 'Couchbase Server Zip'

Change-Id: Iaa02a81e05dcc39ce3338f530ebd6ec4a2fab050
Reviewed-on: http://review.couchbase.org/8631
Reviewed-by: Farshid Ghods <farshid.ghods@gmail.com>
Tested-by: Farshid Ghods <farshid.ghods@gmail.com>