Explicitly start epmd.
We were starting an erl node so epmd would start as a side-effect instead
of just explicitly starting epmd.  That's kind of silly as the erl thing
would fail to start epmd when epmd only was available via a directory with
a space in the name.

Also touched another config file that was referenced.  Don't like mixing
commits, but these two things got the build working.

Change-Id: Idf054f26a4842e6668ab52dfbb1f79c4a431274e
Reviewed-on: http://review.couchbase.org/8736
Tested-by: Farshid Ghods <farshid.ghods@gmail.com>
Reviewed-by: Farshid Ghods <farshid.ghods@gmail.com>