Fixed the library importer/fixer-upper
The code in install-couchbase.sh that copied external libraries into
lib/ and fixed up the imports didn't handle '@loader_path', which is
now used by the current version of the ICU libraries. This caused
dyld errors at launch time.

Fixed this by re-implementing the library fixup as a Ruby script that
recursively traverses dependency chains; that way when it's fixing up
a library that has an '@loader_path' import it remembers where that
library originally came from so it can look there for the imported
library.

I also added a line to the build script to delete the entire
couchbase-core directory before assembling it, so that stuff from
previous builds doesn't get left behind in it.

Change-Id: I5b6732344854c60782148c046f7bfbdc584ffe02
Reviewed-on: http://review.couchbase.org/15880
Reviewed-by: Farshid Ghods <farshid@couchbase.com>
Tested-by: Farshid Ghods <farshid@couchbase.com>