MB-5922: Add README.txt to Zip of Mac app
The zip archive created by the Mac app's "Couchbase Server Zip"
target didn't contain a README file. This patch adds one.

The content of the existing README contained statements that applied
only to the disk image (i.e. the warning against running from the
read-only DMG filesystem) so I created a copy of the README.txt and
tailored it for the Zip file.

To create an archive with multiple items in it I had to switch from
using 'ditto' to 'zip', but this shouldn't cause any noticeable
change.

Change-Id: I173321cd28f1d1a196e17fa9018b7de5e1a99847
Reviewed-on: http://review.couchbase.org/18416
Reviewed-by: Farshid Ghods <farshid@couchbase.com>
Tested-by: Jens Alfke <jens@couchbase.com>