MB-16435: OS X: Don't set DYLD_LIBRARY_PATH
As of OS X 10.11 (El Capitan) it is no longer possible to set /
override DYLD_LIBRARY_PATH. As such we need to move to embedding
runtime linker search paths in all required binaries.

Given we can't rely on this env var anymore, remove setting it in the
startup scripts to (1) ensure we don't incorrectly rely on it and (2)
help in validating that all uses are identified and fixed.

Change-Id: I602a6f014fe4846a93ef80bcd7e9da05fb600a55
Reviewed-on: http://review.couchbase.org/55762
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>