MB-13518: change couchdbx-app to read from ns_server's stderr
couchdbx-app has long read from ns_server's stdout to determine
the point at which the admin console (menelaus) is up and ready
to serve content at which point it opens localhost:8091.

The only problem with this is that ns_server has been console
logging to stderr for what looks to be a similarly long time.
So I'm not entirely sure when this worked and when it broke.

No matter, the fix is for couchdbx to read from ns_server's
stderr and the admin console opens as it should. Shutting down
Couchbase from the menu also seems good.

Also changed: I set the stderr log level to be info so that the
OSX specific Couchbase.log (living in ~/Library/Logs) doesn't
grow too rapidly. This is done by setting the sterr loglevel to
info.

Change-Id: Ic6760123b854110c634824077237949748608bca
Reviewed-on: http://review.couchbase.org/56482
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jens Alfke <jens@couchbase.com>