MB-14849 - Couchbase Server does not start at login on MacOS
The code written by Dustin back in March, 2011, was a hack to begin
with, and the approach it used no longer worked for making Couchbase
Server start automatically on login. Jens looked into this, and found
that there is currently no non-deprecated approach to this, but there
is deprecated code that actually works:

https://github.com/hlung/LaunchAtLoginController/

This change removes the non-functional code in LoginItemManeger, and
replaces it with calls to LaunchAtLoginController. The new code does
work correctly in adding the Couchbase Server.app to the login items
on the mac.

I discovered one other problem while working on this: typing 'make'
alone in the top level directory would fail with a crytic error
messoage. It turns out that one needs to use either: "make
LICENSE=LICENSE-enterprise.txt" or "make
LICENSE=LICENSE-community.txt". I added a comment to the makefile to
indicate as much, as well as a rule to output an error message if
LICENSE is not specified.

Change-Id: Iea33f282f2b45ab6b63814c55c104e38ab60190f
Reviewed-on: http://review.couchbase.org/63151
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Jens Alfke <jens@couchbase.com>