MB-19760: use LSSharedFileList API available on 10.9
In a previous change we introduced the use of LSSharedFileListItemCopyResolvedURL
a convenience function that resolves and returns the URL associated with a file
list item. The problem with this API is that it was introduced in 10.10
which means it's going to break for Mavericks users. From the Apple API docs:

extern CFURLRef
LSSharedFileListItemCopyResolvedURL(
  LSSharedFileListItemRef           inItem,
    LSSharedFileListResolutionFlags   inFlags,
      CFErrorRef *                      outError)
      __OSX_AVAILABLE_STARTING(__MAC_10_10, __IPHONE_NA);

Solution is to use a deprecated, but at least available on 10.9 API:
LSSharedFileListItemResolve. The API doc is as follows for this fucnction:

extern OSStatus
LSSharedFileListItemResolve(
  LSSharedFileListItemRef           inItem,
    LSSharedFileListResolutionFlags   inFlags,
      CFURLRef *                        outURL,
        FSRef *                           outRef)
        __OSX_AVAILABLE_BUT_DEPRECATED_MSG(__MAC_10_5, __MAC_10_10, __IPHONE_NA, __IPHONE_NA,
        "Use LSSharedFileListItemCopyResolvedURL instead.");

We should support Mavericks for now and we can always move to the non-deprecated
version in 5.0.

Change-Id: Ibd24bb034bf4e6d9d218369f2891230cb54b31ac
Reviewed-on: http://review.couchbase.org/64553
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>