MB-19800 - offer defaults for LICENSE and PRODUCT_VERSION in manual builds.
In the MAC build, build errors will occur if LICENSE is not specified.

In addition, runtime errors will happen on MacOS 10.9 (Mavericks) if
PRODUCT_VERSION is not specified.

This change to the makefile provide default values for LICENSE
("LICENSE-community.txt") and PRODUCT_VERSION ("0.0.0-1000").

It should only affect manual builds, as we expect the automated
builds to provide values for LICENSE and PRODUCT_VERSION.

Change-Id: I8e6542b16b65ead8801a5d76c04b48e165a836b5
Reviewed-on: http://review.couchbase.org/64868
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>