MB-22945: sleep 1 before trying to stop epmd
When a user quits the couchdbx-app, the babysitter is stopped,
and immediately epmd is also stopped. However, there's a race between
sending the signal to epmd and the sub-process Erlang VMs disconnecting
from epmd. This results in epmd almost never being stopped. With the sleep
epmd appears to be almost always stopped.

Change-Id: I14f4fda9f59aa9b085f7670b72dc71d116e76aa0
Reviewed-on: http://review.couchbase.org/74113
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Jens Alfke <jens@couchbase.com>