MB-31987: Update MacOS install script to call...
...new helper function to determine the distribution type to use.

Previously, the distribution type to be passed while starting the
babysitter VM was determined based on the content of static_config
file. The server will now be configured using a REST API and a new
helper function has been implemented to determine the distribution
type. The change here is to update the startup script on MacOS to
use this new API.

Change-Id: I37b82dbb4f73325a8f087e107aee79dfe5fe8661
Reviewed-on: http://review.couchbase.org/101671
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>