MB-34266 - Mac app doesn't quit the first time.
For a long time the Mac Couchbase app wouldn't quit the first time you
selected "Quit" from the menubar menu - instead you had to click the
icon again to get the application to stop. It appears that the code
was using an obsolete call to stop the app, this change makes it use a
more modern call.

Change-Id: I7562d9cff0460af9ad1e7cccffba98a14bc34e3c
Reviewed-on: http://review.couchbase.org/109534
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>