MB-30742 - Couchbase Server taking 130% CPU on Mac Mojave
Couchbase Server on the Mac has a wrapper application which starts and
stops the underlying database process, provides the toolbar menu, and
a small number of functions such as checking for updates,
automatically launching the server on login, etc. While the app is
quite simple, the bug reports that on occasion, the wrapper app (as
opposed to the actual database) starts using 130% CPU, and the
database becomes less responsive.

There isn't anything obvious in the wrapper app that could use that
much CPU. Sampling the application shows only the 'dataReady'
function, and many many MacOS system frameworks dealing with file
notifications. After much debugging it appears that the problem lies
in handling output from the database process.

In particular, the wrapper app launches the 'start-server.sh' script,
sending its output to a file handle that gets monitored by the
'dataReady' function. The output is checked for the string "Couchbase
Server has started on web port 8091". When that string is seen, if the
user had selected the "Launch Admin Console at Start" feature, then a
browser is launched pointing at the admin console. Even once this
happens, the wrapper app continues scanning for that string in the
output of ns_server indefinitely.

In MacOS Mojave, there appears to be some issue with dealing with the
file handler used for monitoring the 'start-server.sh' output. Every
few days, it gets into a bad state and starts using 130% of the CPU
for half a day or so, then behavior returns to normal.

The fix that has worked in testing is to directly pipe the output of
'start-server.sh' to ~/Library/Logs/CouchbaseServer.log. Since the
output is no longer being monitored, the "Launch Admin Console at
Start" feature is instead driven by a periodic check of the REST
endpoint /whoami.

Change-Id: I16a232fbd6421769cdab77054e3ea3bd90fcff83
Reviewed-on: http://review.couchbase.org/110682
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>
Tested-by: Dave Finlay <dave.finlay@couchbase.com>