Make changes to the Mac build to permit headless DMG file creation.
With the upgrade to MacOS Mojave, sandboxing prevents a script from
using AppleScript to control the Finder unless the user responds
affirmatively to a permission dialog. The script we had been using for
creating nice DMG files, create-dmg, using AppleScript to perform
Finder operations on the disk image. Since there is no user in the
loop for builds, this led to error #610 in the build logs (and a very
plain looking DMG).

There exists a new solution to creating DMG files without user
intervention, basing a new DMG file on an existing template. This
change to the Mac build includes the template. A separate change to
the create-dmg.sh script used by Jenkins will use that template when
creating the DMG file.

Change-Id: I4ba996467455c3e556ccda87ab1821b59474e531
Reviewed-on: http://review.couchbase.org/116169
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>