Disable "prevent overlapping partitions" in erl dist
Disable the new check in erlang 25 that "prevents overlapping
partitions". It is tied to the new facility of the "global" module
that gives a few features such as name registration and simple locking
across a cluster. We do not use this module, but instead have
implemented this ourselves. Perhaps this upgraded "global" module
could be used in place of our custom solution, but in the meantime
this causes problems while nodes are attempting to join together in
some situations. It has also caused some problems with remote shell
connections. Changing this setting puts this feature back to the way
it was in erlang 24 so we can decide at another time if it's worth
using the new changes from the global module.

Here are some link(s) that explains the new feature and why this is
problematic:
 - https://www.erlang.org/patches/otp-25.0
 - https://github.com/erlang/otp/pull/5611
 - https://github.com/erlang/otp/pull/5687
 - https://stackoverflow.com/q/73567169
 - https://stackoverflow.com/a/73578740

Change-Id: I96b7bf67a6a0a41230f5b10e5c80dab1005e5be8
Reviewed-on: https://review.couchbase.org/c/couchdbx-app/+/187155
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>