MB-66105: fix mac installer profile loading
The mac installer hasn't been updated to properly load the
default/provisioned profiles especially in a way that will work with the
new columnar changes. This patch fixes that by checking for the files
existence (from CMake) and then it'll copy those if they exist.

This means in columnar where those profiles aren't produced, we won't
accidentally include them.

This wasn't that big of a deal because nobody used provisioned on mac
and there weren't many keys being used in the default profile.

Change-Id: If66474aa3bffe6b2484733183cc75dfd09c5e7a9
Reviewed-on: https://review.couchbase.org/c/couchdbx-app/+/225827
Reviewed-by: Michael Blow <michael.blow@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>